constants.php-sample: Removing unnecessary logic
This commit removes logic that does not belong in a config
file. Using the same constats.php file for multiple instances
of the site should be an extremely rare scenario, and thus
the logic for this should not be in the example file.
Anyone who needs this functionality should be skilled enough
to reimplement it in their own constants.php file.

Added bonus: This should save us a couple of CPU cycles.

Signed-off-by: Martin Svangren <martin@masv.net>