constants.php-sample: Adding comments
This comment adds a few comments to the example
configuration file. The available parameters may seem
obvious to experienced coders, but could be quite
confusing to others.

Some of the comments are more directed at the core
development team than contributing developers. Some
config settings are duplicated in variables and
constants, which increases clutter in the file as
well as wastes CPU and memory cycles. At some point,
this file should get an overhaul and be cleaned up
to contain the bare necessities needed to get an
instance of the code up and running.

Signed-off-by: Martin Svangren <martin@masv.net>